# Optimize BaseParser#unnormalize method