@naitoh 2024-06-24
After removing `rv = string.gsub( /\r\n?/,"\n")`, the effect is not so great...