@kou 2024-06-25
Thanks.

```ruby
if string.include?("\r")
  rv = string.gsub( /\r\n?/, "\n" )
else
  rv = string.dup
end
```

may be faster.